third_party/blink/web_tests/html/sections/hgroup-element-expected.txt

Various tests for the hgroup element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

<hgroup> closes <p>:
PASS hgroup1.parentNode.nodeName == "p" is false
<p> does not close <hgroup>:
PASS p1.parentNode.nodeName is "HGROUP"
<hgroup> can be nested inside <hgroup>:
PASS hgroup3.parentNode.id is "hgroup2"
Residual style:
PASS getWeight("hgroup4") is "700"
PASS getWeight("span1") is "700"
FormatBlock:
PASS document.getElementById("span2").parentNode.nodeName is "HGROUP"
PASS successfullyParsed is true

TEST COMPLETE
